+++ scripts/clean.js
@@ -1,5 +1,5 @@
 const sh = require('shelljs');
-const upath = require('upath');
+const upath = require('@ewizardjs/upath');
 
 const destPath = upath.resolve(upath.dirname(__filename), '../dist');