+++ scripts/render-assets.js
@@ -1,6 +1,6 @@
 'use strict';
 const fs = require('fs');
-const upath = require('upath');
+const upath = require('@ewizardjs/upath');
 const sh = require('shelljs');
 
 module.exports = function renderAssets() {