+++ scripts/render-scripts.js
@@ -1,7 +1,7 @@
 'use strict';
 const fs = require('fs');
 const packageJSON = require('../package.json');
-const upath = require('upath');
+const upath = require('@ewizardjs/upath');
 const sh = require('shelljs');
 
 module.exports = function renderScripts() {