+++ scripts/start.js
@@ -1,5 +1,5 @@
 const concurrently = require('concurrently');
-const upath = require('upath');
+const upath = require('@ewizardjs/upath');
 
 const browserSyncPath = upath.resolve(upath.dirname(__filename), '../node_modules/.bin/browser-sync');